azure_function/requirements.txt
azure-functions
azure-identity
azure-mgmt-datafactory
azure-mgmt-monitor
google-generativeai
jinja2
python-dotenv
